refactor(ui): extract project work item detail
